Continue to refine challenge descriptions
Change 5 more code challenges, add README files with the descriptions
