+++ src/ransom-note/packageransomnote/Main.java
@@ -1,38 +1,9 @@
-package packageRansomNote;
+package packageransomnote;
 
 import java.util.HashMap;
 
 public class Main {
 
-//	Challenge:
-//	
-//	Given two strings ransomNote and magazine, return true if ransomNote can be constructed by using the letters 
-//	from magazine and false otherwise.
-//
-//	Each letter in magazine can only be used once in ransomNote.
-//
-  
-//	Example 1:
-//
-//	Input: ransomNote = "a", magazine = "b"
-//	Output: false
-  
-//	Example 2:
-//
-//	Input: ransomNote = "aa", magazine = "ab"
-//	Output: false
-  
-//	Example 3:
-//
-//	Input: ransomNote = "aa", magazine = "aab"
-//	Output: true
-//			 
-//
-//	Constraints:
-//
-//	1 <= ransomNote.length, magazine.length <= 105
-//	ransomNote and magazine consist of lowercase English letters.
-  
   public static void main(String[] args) {
 
     String ransomNote = "aa", magazine = "aab";
