+++ src/single-number/packagesinglenumber/Main.java
@@ -1,4 +1,4 @@
-package packageSingleNumber;
+package packagesinglenumber;
 
 import java.util.Arrays;
 import java.util.HashMap;
@@ -8,35 +8,6 @@ import java.util.Optional;
 import java.util.stream.IntStream;
 
 public class Main {
-
-//	Challenge:
-//	
-//	Given a non-empty array of integers nums, every element appears twice except for one. Find that single one.
-//
-//	You must implement a solution with a linear runtime complexity and use only constant extra space.
-//
-//	 
-//	Example 1:
-//
-//	Input: nums = [2,2,1]
-//	Output: 1
-//	
-//	Example 2:
-//
-//	Input: nums = [4,1,2,1,2]
-//	Output: 4
-//	
-//	Example 3:
-//
-//	Input: nums = [1]
-//	Output: 1
-//	 
-//
-//	Constraints:
-//
-//	1 <= nums.length <= 3 * 104
-//	-3 * 104 <= nums[i] <= 3 * 104
-//	Each element in the array appears twice except for one element which appears only once.
 
   public static void main(String[] args) {
